packages/agentic-workflows/pyproject.toml
[project]
name = "zeroshot-agentic-workflows"
version = "0.1.0"
description = "Framework-agnostic agent workflow building blocks for Zeroshot Python packages."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
  "zeroshot-commons==0.1.0",
  "openai-agents>=0.0.7",
  "pyyaml>=6.0",
]

[tool.uv.sources]
zeroshot-commons = { workspace = true }

[build-system]
requires = ["uv_build>=0.11.8,<0.12"]
build-backend = "uv_build"
